backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.3
reportlab==4.0.9
